+++ CHANGELOG.md
@@ -10,6 +10,12 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Support mdbook-linkcheck.
+- Support mdbook on Windows.
+- Support wasmtime on Windows.
+- Support nextest on Linux (musl).
+- Fix installation failure when `$CARGO_HOME/bin` or `$HOME/.cargo/bin` is not available.
+
 ## [1.5.11] - 2022-07-25
 
 - Fix cargo-binstall installation on macOS and Windows.